fix test, jslint ignores missing moment
add unit test dependencies to fix test

moment is in the "old school" of js libraries that imports something
into the global namespace. This is outside of what a "hygenic" module
system, which jslint deals with, should do. Thus, we have to tell our
jslinter that there is a global symbol moment, "magically" appearing in
our global namespace even though we're never setting it explicitly.